requirements.txt
Django==4.2.29
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf_yasg==1.21.15
python-dotenv==1.2.1
mysqlclient==2.2.7
gunicorn==21.2.0